Revert "database sequence"
This reverts commit a8e4809178fa8dcc7eb1c441daf4204841bb8992.

+++ logic/src/main/java/application/Repository/UserRepository.java
@@ -10,5 +10,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByEmail(String username);
 
-    List<User> findByCompanyId(long company);
+    List<User> findByCompanyId(int company);
 }
